Save device name instead of device object
Save the device name instead of the device object. There was some
issues when unplugging a device and plugging it back in that it
would not cycle as excpected. Now it does a device lookup on each
swap, add, and del. So far the overhead does not appear to impact
performance, but I don't have many devices I swap between.